게시물 수정, 삭제 권한체크, 로그인체크, jsp include 사용법

+++ src/main/webapp/jsp/article/detail.jsp
@@ -6,6 +6,7 @@
 
 <%
 	Map<String, Object> articleRow = (Map) request.getAttribute("articleRow");
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 %>
 
 <html>
@@ -23,10 +24,16 @@
 	<div>내용 : <%= (String) articleRow.get("body") %></div>
 	<div>
 		<a href="list">목록</a>
-		<a href="modify?id=<%= (int) articleRow.get("id") %>">수정</a>
-		<a href="doDelete?id=<%= (int) articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		<%
+		if (loginedMemberId != -1 && (int) articleRow.get("memberId") == loginedMemberId) {
+		%>
+			<a href="modify?id=<%= (int) articleRow.get("id") %>">수정</a>
+			<a href="doDelete?id=<%= (int) articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		<%
+		}
+		%>
 	</div>
+	
 </body>
 </html>
-
